% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ascus}
\alias{ascus}
\title{ASCUS dataset}
\format{A data frame of 20 observations and six variables:
\describe{
 \item{StudyID}{Study identifier}
 \item{Test}{Type of diagnostic test}
 \item{TP}{Number of true positives}
 \item{FN}{number of false negatives}
 \item{TN}{number of true negatives}
 \item{FP}{number of false positives}
}}
\usage{
data(ascus)
}
\description{
Arbyn et al. (2013) performed a Cochrane review on the accuracy of human papillomavirus
testing (HC2)and repeat cytology (RepC) to triage of women with an equivocal Pap smear to diagnose cervical precancer.
}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
\references{
{Arbyn M, Roelens J, Simoens C, Buntinx F, Paraskevaidis E, Martin-Hirsch PPL, Prendiville W (2013). Human Papillomavirus Testing Versus Repeat
Cytology for Triage of Minor Cytological Cervical Lesions." Cochrane Database of Systematic Reviews, pp. 31-201.}
}

