% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cdtafit-class}
\alias{cdtafit-class}
\title{Class cdtafit}
\description{
A cdtafit class in the CopulaDTA package.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data-frame with no missing values containg TP, TN, FP, FN, 'SID' and co-varaiables(if necessary).}

\item{\code{SID}}{A string indicating the name of the column with the study identifier.}

\item{\code{copula}}{Copula function, 'fgm', 'gauss', 'c90', '270', or 'frank'.}

\item{\code{modelargs}}{List containing control parameters for the prior distributions.}

\item{\code{fit}}{An object of class stanfit returned by the function sampling.}
}}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
\seealso{
\link{fit}

Other cdta: \code{\link{cdtamodel-class}}
}

