% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cdtamodel-class}
\alias{cdtamodel-class}
\title{Class cdtamodel}
\description{
A cdtamodel class in the CopulaDTA package.
}
\section{Slots}{

\describe{
\item{\code{copula}}{Copula function, 'fgm', 'gauss', 'c90', '270', or 'frank'.}

\item{\code{modelcode}}{Character with the model code as returned by the model function}

\item{\code{modelargs}}{List containing control parameters for the prior distributions}
}}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
\seealso{
\link{cdtamodel}

Other cdta: \code{\link{cdtafit-class}}
}

