% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ft}
\alias{ft}
\title{Compute transform omega to ktau.}
\usage{
ft(t)
}
\arguments{
\item{theta}{correlation parameter(s) from the frank copula function.}
}
\description{
Compute transform omega to ktau.
}
\author{
Victoria N Nyaga
}
\keyword{internal}
