% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{omega.to.ktau}
\alias{omega.to.ktau}
\title{Compute transform omega to ktau.}
\usage{
omega.to.ktau(omega)
}
\arguments{
\item{theta}{correlation parameter(s) from the frank copula function.}
}
\description{
Compute transform omega to ktau.
}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
\keyword{internal}

