% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,cdtamodel-method}
\alias{show,cdtafit-method}
\alias{show,cdtamodel-method}
\title{Print a cdtamodel object.}
\usage{
\S4method{show}{cdtamodel}(object)

\S4method{show}{cdtafit}(object)
}
\arguments{
\item{object}{A cdtamodel object returned by \link{cdtamodel} function.}
}
\description{
Print a cdtamodel object.

Print a cdtafit object.
}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
\keyword{internal}

