% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telomerase}
\alias{telomerase}
\title{Telomerase dataset}
\format{A data frame of ten observations and five variables:
\describe{
 \item{ID}{Study identifier}
 \item{TP}{Number of true positives}
 \item{FN}{number of false negatives}
 \item{TN}{number of true negatives}
 \item{FP}{number of false positives}
}}
\usage{
data(telomerase)
}
\description{
Glas et al. (2003) systematically reviewed the sensitivity and specificity of cytology and other
markers including telomerase for primary diagnosis of bladder cancer.
}
\references{
{Glas AS, Roos D, Deutekom M, Zwinderman AH, Bossuyt PMM, Kurth KH (2003). Tumor Markers in the Diagnosis of Primary Bladder Cancer.
A Systematic Review. The Journal of Urology, 169(6), 1975-1982.}
}

