% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{tracecopula}
\alias{tracecopula}
\title{Trace plot using ggplot2.}
\usage{
tracecopula(object, ...)
}
\arguments{
\item{object}{An object from \link{fitcopula}.}

\item{...}{additional options. See \link[rstan]{stan_trace} for more details.}
}
\value{
A ggplot trace plot of the parameters of the models mean structure.
}
\description{
Trace plot using ggplot2.
}
\examples{
\dontrun{
fit <- fitcopula(data=ascus,
         SID = "StudyID",
         formula.se= StudyID ~ Test,
         cores=3,
         seed=3,
         copula="fgm")

tracecopula(fit)
}
}
\author{
Victoria N Nyaga
}

