% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{traceplot}
\alias{traceplot}
\alias{traceplot,cdtafit-method}
\title{Traceplots for cdtafit objects.}
\usage{
traceplot(object, ...)

\S4method{traceplot}{cdtafit}(object, ...)
}
\arguments{
\item{object}{A cdtafit object from \link{fit}}

\item{...}{Extra optional arguments as defined in \link[rstan]{stan_trace}.}
}
\description{
Traceplots for cdtafit objects.
}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}

