% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{waic}
\alias{waic}
\title{Compute log pointwise predictive density, effective number of parameters and WAIC.}
\usage{
waic(model)
}
\arguments{
\item{model}{stanfit object}
}
\description{
Compute log pointwise predictive density, effective number of parameters and WAIC.
}
\author{
Victoria N Nyaga
}
\keyword{internal}
