% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{cdtafit-class}
\alias{cdtafit-class}
\title{Class cdtafit}
\description{
A cdtafit class in the CopulaDTA package.
}
\section{Slots}{

\describe{
\item{\code{data}}{a data-frame with no missing values containing TP, TN, FP, FN, 'SID' and co-variables(if necessary).}

\item{\code{SID}}{A string indicating the name of the column with the study identifier.}

\item{\code{copula}}{copula function, 'fgm', 'gauss', 'c90', '270', or 'frank'.}

\item{\code{modelargs}}{list containing control parameters for the prior distributions.}

\item{\code{fit}}{an object of class stanfit returned by the function sampling.}
}}

\seealso{
\link{fit}

Other cdta: 
\code{\link{cdtamodel-class}}
}
\author{
Victoria N Nyaga \email{victoria.nyaga@outlook.com}
}
\concept{cdta}
