% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,cdtamodel-method}
\alias{show,cdtamodel-method}
\alias{show,cdtafit-method}
\title{A function to print the model.}
\usage{
\S4method{show}{cdtamodel}(object)

\S4method{show}{cdtafit}(object)
}
\arguments{
\item{object}{A cdtamodel object returned by \link{cdtamodel} function.}
}
\description{
A function to print the model.

A function to print the results.
}
\keyword{internal}
