% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceplot.R
\name{traceplot.cdtafit}
\alias{traceplot.cdtafit}
\title{Trace plot using ggplot2.}
\usage{
traceplot.cdtafit(x, ...)
}
\arguments{
\item{x}{An cdtafit object from \link{fit}.}

\item{...}{additional options. See \link[rstan]{stan_trace} for more details.}
}
\value{
A ggplot trace plot of the parameters of the models mean structure.
}
\description{
Trace plot using ggplot2.
}
\examples{
\dontrun{
fit1 <- fit(model1,
                SID='ID',
                data=telomerase,
                iter=2000,
                warmup=1000,
                thin=1,
                seed=3)

traceplot(fit1)
}
}
\author{
Victoria N Nyaga
}

