% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{traceplot}
\alias{traceplot}
\alias{traceplot,cdtafit-method}
\title{A function to produce traceplots.}
\usage{
traceplot(object, ...)

\S4method{traceplot}{cdtafit}(object, ...)
}
\arguments{
\item{object}{A cdtafit object from \link{fit}}

\item{...}{Extra optional arguments as defined in \link[rstan]{stan_trace}.}
}
\description{
A function to produce traceplots.

A function to produce traceplots.
}
